# scripts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Contacts API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/followups": {
      "get": {
        "tags": [
          "Followups"
        ],
        "summary": "List Followups",
        "operationId": "list_followups_followups_get",
        "parameters": [
          {
            "name": "contactId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contactid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FollowUpItem"
                  },
                  "title": "Response List Followups Followups Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Followups"
        ],
        "summary": "Create Followup",
        "operationId": "create_followup_followups_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FollowUpCreate": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "contactId": {
            "type": "string",
            "title": "Contactid"
          },
          "channel": {
            "type": "string",
            "enum": [
              "call",
              "email",
              "whatsapp",
              "meeting"
            ],
            "title": "Channel"
          },
          "dueAt": {
            "type": "string",
            "title": "Dueat"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "contactId",
          "channel",
          "dueAt"
        ],
        "title": "FollowUpCreate"
      },
...
      "FollowUpItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "contactId": {
            "type": "string",
            "title": "Contactid"
          },
          "channel": {
            "type": "string",
            "enum": [
              "call",
              "email",
              "whatsapp",
              "meeting"
            ],
            "title": "Channel"
          },
          "dueAt": {
            "type": "string",
            "title": "Dueat"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "contactId",
          "channel",
          "dueAt",
          "createdAt"
        ],
        "title": "FollowUpItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}